Refactor: Move common generators to common folder for csharp
