--- common/src/codegen/playwright-charp-common/basePlaywrightDotnetTemplatesProvider.ts
@@ -1,2 +1,0 @@
-/** Base class for PlaywrightDotnetTemplatesProvider */
-export class BasePlaywrightDotnetTemplatesProvider {}
