+++ common/src/codegen/playwright-charp-common/templates/BasePageFile.hbs
