+++ common/src/codegen/playwright-charp-common/templates/TestCaseBaseFile.hbs
