+++ common/src/codegen/playwright-charp-common/templates/TestRoutineClass.hbs
