+++ common/src/codegen/playwright-csharp-nunit/playwrightCsharpNunitCodeGen.ts
@@ -1,43 +1,31 @@
 import { EOL } from "os";
 import path from "path";
-import {
-  ISourceProjectMetadata,
-  ITestCase,
-  ITestRoutine,
-  ITestSuite,
-  LocatorType,
-  StandardFolder,
-  StandardOutputFolder,
-} from "../../file-defs";
-import { IPage } from "../../file-defs/pageFile";
+import { ISourceProjectMetadata, ITestCase, ITestSuite, StandardFolder, StandardOutputFolder } from "../../file-defs";
 import { StandardOutputFile } from "../../file-defs/standardOutputFile";
 import { ICodeGen, WriteFileFn } from "../types";
-import { hasPlaceholder, upperCaseFirstChar } from "../utils/stringUtils";
 import { NunitProjectMetaGenerator } from "./nunitProjectMetaGenerator";
-import { PlaywrightCsharpNunitTemplatesProvider } from "./playwrightCsharpNunitTemplatesProvider";
-import { IDataSetInfo } from "../playwright-charp-common/dataSetInfo";
-import { PlaywrightCsharpCodeGen } from "../playwright-charp-common/playwrightCsharpCodeGen";
+import { PlaywrightCsharpNUnitTemplatesProvider } from "./playwrightCsharpNunitTemplatesProvider";
+import { CommonPlaywrightCsharpCodeGen } from "../playwright-charp-common/commonPlaywrightCsharpCodeGen";
 import { IOutputProjectMetadataGenerator } from "../playwright-charp-common/outputProjectMetadataProcessor";
-import generateDatasetInfos from "../playwright-charp-common/generateDatasetInfos";
 import { IPlaywrightCsharpTemplatesProvider } from "../playwright-charp-common/playwrightCsharpTemplatesProvider";
 import { createOutputProjectMetadata } from "../codegenOutputProjectMeta";
 import { addIndent } from "../utils/codegenUtils";
 
-export class PlaywrightCsharpNunitCodeGen extends PlaywrightCsharpCodeGen implements ICodeGen {
+export class PlaywrightCsharpNunitCodeGen extends CommonPlaywrightCsharpCodeGen implements ICodeGen {
+  private _templateProvider: PlaywrightCsharpNUnitTemplatesProvider;
+
   constructor(projMeta: ISourceProjectMetadata) {
     super(projMeta);
-  }
 
-  protected override getOutProjMeta(): IOutputProjectMetadataGenerator {
-    return new NunitProjectMetaGenerator(this._projMeta);
-  }
-
-  protected override getTemplateProvider(): IPlaywrightCsharpTemplatesProvider {
-    return new PlaywrightCsharpNunitTemplatesProvider(
+    this._templateProvider = new PlaywrightCsharpNUnitTemplatesProvider(
       path.join(this._rmprojFile.folderPath, StandardFolder.CustomCode, "templates"),
       this._rmprojFile.content.indent,
       this._rmprojFile.content.indentSize
     );
+  }
+
+  protected override getOutProjMeta(): IOutputProjectMetadataGenerator {
+    return new NunitProjectMetaGenerator(this._projMeta);
   }
 
   async generateCode(full: boolean, writeFile: (path: string, content: string) => Promise<void>): Promise<string> {
@@ -45,10 +33,11 @@ export class PlaywrightCsharpNunitCodeGen extends PlaywrightCsharpCodeGen implem
     await this.generateEnvironmentSetterScripts(writeFile);
 
     await this.generatePageFiles(writeFile);
-    await this.generateCaseFiles(writeFile);
+    await this.generateTestCaseFiles(writeFile);
     await this.generateRoutineFiles(writeFile);
-    await this.generateSuiteFiles(writeFile);
+    await this.generateTestSuiteFiles(writeFile);
     await this.generateSupportFiles(writeFile);
+    await this.generateNUnitSupportFiles(writeFile);
 
     if (full) {
       await this.generateProjectFiles(writeFile);
@@ -59,52 +48,7 @@ export class PlaywrightCsharpNunitCodeGen extends PlaywrightCsharpCodeGen implem
     return "";
   }
 
-  async generateRoutineFiles(writeFile: WriteFileFn) {
-    // Filename: TestRoutines/{TestRoutineName}.cs
-    for (let { content: testRoutine } of this._projMeta.testRoutines) {
-      const datasets: IDataSetInfo[] = generateDatasetInfos(testRoutine);
-      const testRoutinesClasses: string[] = [];
-
-      // For each dataset, we generate a separate routine class
-      for (let dataset of datasets) {
-        let testRoutineClass = this.generateTestRoutineClass(
-          testRoutine,
-          this._projMeta.pages.map((p) => p.content),
-          dataset
-        );
-        testRoutinesClasses.push(testRoutineClass);
-      }
-
-      let testRoutineFile = this.generateTestRoutineFile(testRoutine, testRoutinesClasses);
-      let outputFileRelPath = this._outProjMeta.get(testRoutine.id)!.outputFileRelPath;
-      await writeFile(outputFileRelPath, testRoutineFile);
-    }
-  }
-
-  async generateCaseFiles(writeFile: WriteFileFn) {
-    // # Generate TestCase files
-
-    for (let { content: testCase } of this._projMeta.testCases) {
-      let testClassContent = this.generateTestCaseFile(
-        testCase,
-        this._projMeta.pages.map((p) => p.content),
-        this._projMeta.testRoutines.map((p) => p.content)
-      );
-      let outputFileRelPath = this._outProjMeta.get(testCase.id)!.outputFileRelPath;
-
-      await writeFile(outputFileRelPath, testClassContent);
-    }
-  }
-  async generatePageFiles(writeFile: WriteFileFn) {
-    // # Generate Page definition for each page
-    // Filename: Pages/{PageName}.cs
-    for (let page of this._projMeta.pages) {
-      let filePath = this._outProjMeta.get(page.content.id)!.outputFileRelPath;
-      await writeFile(filePath, this.generatePage(page.content));
-    }
-  }
-  async generateSuiteFiles(writeFile: WriteFileFn) {
-    // Filename: TestSuites/{TestClassName}.cs
+  async generateTestSuiteFiles(writeFile: WriteFileFn) {
     for (let testSuite of this._projMeta.testSuites) {
       let fileRelPath = this._outProjMeta.get(testSuite.content.id)!.outputFileRelPath;
       let classContent = this.generateTestSuiteFile(
@@ -115,31 +59,17 @@ export class PlaywrightCsharpNunitCodeGen extends PlaywrightCsharpCodeGen implem
     }
   }
 
-  async generateSupportFiles(writeFile: WriteFileFn) {
-    // Each page definition will be a property
-    // Filename: PageDefinitions.cs
+  async generateNUnitSupportFiles(writeFile: WriteFileFn) {
     await writeFile(
-      `${StandardOutputFile.PageDefinitions}${this._outputFileExt}`,
-      this.generatePageDefinitions(this._projMeta.pages.map((p) => p.content))
-    );
-
-    // Filename: Support/TestCaseBase.cs
-    await writeFile(
-      `${StandardOutputFolder.Support}/${StandardOutputFile.TestCaseBase}${this._outputFileExt}`,
-      this._templateProvider.getTestCaseBase(this._rmprojFile.content.rootNamespace)
-    );
-
-    // Filename: Support/TestSuiteBase.cs
-    await writeFile(
-      `${StandardOutputFolder.Support}/${StandardOutputFile.TestSuiteBase}${this._outputFileExt}`,
+      `${StandardOutputFolder.NUnitSupport}/${StandardOutputFile.TestSuiteBase}${this._outputFileExt}`,
       this._templateProvider.getTestSuiteBase(
         this._rmprojFile.content.rootNamespace,
         this._rmprojFile.content.testIdAttributeName
       )
     );
   }
+
   async generateMetaFiles(writeFile: WriteFileFn) {
-    // Write output project metadata
     const outputProjectMetadata = await createOutputProjectMetadata(this._rmprojFile);
     await writeFile(StandardOutputFile.MetaData, JSON.stringify(outputProjectMetadata, null, 2));
   }
@@ -154,81 +84,6 @@ export class PlaywrightCsharpNunitCodeGen extends PlaywrightCsharpCodeGen implem
       this._templateProvider.getUsings(this._rmprojFile.content.rootNamespace, this._projMeta.testRoutines.length > 0)
     );
     await writeFile(`${StandardOutputFile.RunSettings}`, this._templateProvider.getRunSettings());
-  }
-
-  generatePageDefinitions(pages: IPage[]): string {
-    let usingDirectives: string[] = [];
-    for (let page of pages) {
-      let pageNamespace = this._outProjMeta.get(page.id)!.outputFileFullNamespace;
-      let usingDirective = `using ${pageNamespace};`;
-      if (!usingDirectives.includes(usingDirective)) {
-        usingDirectives.push(usingDirective);
-      }
-    }
-
-    let usings = usingDirectives.join(EOL);
-
-    let pagesDeclarationItems = [];
-    for (let page of pages) {
-      let pageName = upperCaseFirstChar(this._outProjMeta.get(page.id)!.outputFileClassName);
-      pagesDeclarationItems.push(`public ${pageName} ${pageName} { get; private set; }`);
-    }
-    let pagesDeclarations = pagesDeclarationItems.join(EOL);
-    pagesDeclarations = addIndent(pagesDeclarations, this._indentString);
-    //
-    // Build constructor body
-    // Example:
-    // this.LoginPage = new LoginPage(this);
-    //
-    let pageInitItems = [];
-    for (let page of pages) {
-      let pageName = upperCaseFirstChar(this._outProjMeta.get(page.id)!.outputFileClassName);
-      pageInitItems.push(`${pageName} = new ${pageName}(page);`);
-    }
-
-    let pageInits = pageInitItems.join(EOL);
-    pageInits = addIndent(pageInits, this._indentString, 2);
-
-    return this._templateProvider.getPageDefinitions(usings, this._rootNamespace, pagesDeclarations, pageInits);
-  }
-
-  private generatePage(page: IPage): string {
-    let pageItems = [];
-
-    for (let element of page.elements) {
-      if (element.type === "pageElement") {
-        pageItems.push(
-          this._templateProvider.getLocator({
-            elementName: upperCaseFirstChar(element.name!),
-            locatorStr: element.locator || "",
-            locatorType: element.findBy!,
-            description: element.description!,
-            hasParams: hasPlaceholder(element.locator!),
-            returnedLocatorType:
-              element.findBy! === LocatorType.IFrame ||
-              element.findBy! === LocatorType.IFrameId ||
-              element.findBy! === LocatorType.IFrameName
-                ? "IFrameLocator"
-                : "ILocator",
-          })
-        );
-        continue;
-      }
-      if (element.type === "comment") {
-        pageItems.push(""); // Add an empty line before comment
-        pageItems.push(this._templateProvider.getComment(element.comment!));
-      }
-    }
-
-    let pageBody = pageItems.join(EOL + EOL);
-    pageBody = addIndent(pageBody, this._indentString);
-
-    return this._templateProvider.getPage(
-      this._outProjMeta.get(page.id)!.outputFileFullNamespace,
-      this._outProjMeta.get(page.id)!.outputFileClassName,
-      page.description || "",
-      pageBody
-    );
   }
 
   private generateTestSuiteFile(testSuite: ITestSuite, testcases: ITestCase[]) {
@@ -266,51 +121,5 @@ export class PlaywrightCsharpNunitCodeGen extends PlaywrightCsharpCodeGen implem
       this._outProjMeta.get(testSuite.id)!.outputFileFullNamespace
     );
     return testClass;
-  }
-
-  private generateTestCaseFile(testCase: ITestCase, pages: IPage[], routines: ITestRoutine[]) {
-    const testcaseBody = this.generateTestCaseBody(testCase, pages, routines);
-    const routineUsings = this.generateRoutineUsings(testCase, routines);
-
-    let testFile = this._templateProvider.getTestCaseFile(
-      this._outProjMeta.get(testCase.id)!.outputFileClassName,
-      testCase.description,
-      testcaseBody,
-      this._rootNamespace,
-      this._outProjMeta.get(testCase.id)!.outputFileFullNamespace,
-      routineUsings
-    );
-    return testFile;
-  }
-
-  private generateTestCaseFunction(testCase: ITestCase) {
-    const testcaseName = this._outProjMeta.get(testCase.id)!.outputFileClassName;
-    const testCaseMethod = this._templateProvider.getTestFunction(testcaseName, testCase.description);
-    return testCaseMethod;
-  }
-
-  private generateTestRoutineClass(testRoutine: ITestRoutine, pages: IPage[], datasetInfo: IDataSetInfo) {
-    const testRoutineBody = this.generateTestRoutineBody(testRoutine, pages, datasetInfo);
-
-    // Output class name will be "{testRoutineClassName}{datasetName}";
-    const testRoutineName = this._outProjMeta.get(testRoutine.id)!.outputFileClassName;
-    const finalOutputClassName = `${testRoutineName}${datasetInfo.name}`;
-
-    let routineFileContent = this._templateProvider.getTestRoutineClass(
-      finalOutputClassName,
-      testRoutine.description,
-      testRoutineBody
-    );
-
-    return routineFileContent;
-  }
-
-  private generateTestRoutineFile(testRoutine: ITestRoutine, testRoutineClasses: string[]) {
-    let routineFileContent = this._templateProvider.getTestRoutineFile(
-      this._rootNamespace,
-      this._outProjMeta.get(testRoutine.id)!.outputFileFullNamespace,
-      testRoutineClasses
-    );
-    return routineFileContent;
   }
 }
