+++ common/src/codegen/playwright-csharp-nunit/templates/TestSuiteBaseFile.hbs
@@ -1,4 +1,4 @@
-namespace {{{rootNamespace}}}.Support;
+namespace {{{rootNamespace}}}.NUnitSupport;
 
 [TestFixture]
 public abstract partial class TestSuiteBase : PageTest
