+++ common/src/codegen/playwright-csharp-nunit/templates/UsingsFile.hbs
@@ -2,6 +2,7 @@ global using NUnit.Framework;
 global using Microsoft.Playwright;
 global using Microsoft.Playwright.NUnit;
 global using {{{rootNamespace}}}.Support;
+global using {{{rootNamespace}}}.NUnitSupport;
 global using {{{rootNamespace}}}.Config;
 
 {{#if hasRoutines}}
