+++ common/src/codegen/playwright-csharp-xunit/playwrightCsharpXUnitCodeGen.ts
@@ -16,7 +16,7 @@ import { hasPlaceholder, upperCaseFirstChar } from "../utils/stringUtils";
 import { PlaywrightCsharpXUnitTemplatesProvider } from "./playwrightCsharpXUnitTemplatesProvider";
 import { XUnitProjectMetaGenerator } from "./xunitProjectMetaGenerator";
 import { IDataSetInfo } from "../playwright-charp-common/dataSetInfo";
-import { PlaywrightCsharpCodeGen } from "../playwright-charp-common/playwrightCsharpCodeGen";
+import { CommonPlaywrightCsharpCodeGen } from "../playwright-charp-common/commonPlaywrightCsharpCodeGen";
 import { IPlaywrightCsharpTemplatesProvider } from "../playwright-charp-common/playwrightCsharpTemplatesProvider";
 import { IOutputProjectMetadataGenerator } from "../playwright-charp-common/outputProjectMetadataProcessor";
 import generateDatasetInfos from "../playwright-charp-common/generateDatasetInfos";
@@ -25,21 +25,20 @@ import { addIndent } from "../utils/codegenUtils";
 
 type WriteFileFn = (path: string, content: string) => Promise<void>;
 
-export class PlaywrightCsharpXUnitCodeGen extends PlaywrightCsharpCodeGen implements ICodeGen {
+export class PlaywrightCsharpXUnitCodeGen extends CommonPlaywrightCsharpCodeGen implements ICodeGen {
+  private _templateProvider: PlaywrightCsharpXUnitTemplatesProvider;
+
   constructor(projMeta: ISourceProjectMetadata) {
     super(projMeta);
-  }
-
-  protected override getOutProjMeta(): IOutputProjectMetadataGenerator {
-    return new XUnitProjectMetaGenerator(this._projMeta);
-  }
-
-  protected override getTemplateProvider(): IPlaywrightCsharpTemplatesProvider {
-    return new PlaywrightCsharpXUnitTemplatesProvider(
+    this._templateProvider = new PlaywrightCsharpXUnitTemplatesProvider(
       path.join(this._rmprojFile.folderPath, StandardFolder.CustomCode, "templates"),
       this._rmprojFile.content.indent,
       this._rmprojFile.content.indentSize
     );
+  }
+
+  protected override getOutProjMeta(): IOutputProjectMetadataGenerator {
+    return new XUnitProjectMetaGenerator(this._projMeta);
   }
 
   async generateCode(full: boolean, writeFile: WriteFileFn): Promise<string> {
@@ -52,6 +51,7 @@ export class PlaywrightCsharpXUnitCodeGen extends PlaywrightCsharpCodeGen implem
     await this.generateRoutineFiles(writeFile);
     await this.generateTestSuiteFiles(writeFile);
     await this.generateSupportFiles(writeFile);
+    await this.generateXUnitSupportFiles(writeFile);
 
     if (full) {
       await this.writeProjectFiles(writeFile);
@@ -78,48 +78,14 @@ export class PlaywrightCsharpXUnitCodeGen extends PlaywrightCsharpCodeGen implem
     await writeFile(`${StandardOutputFile.RunSettings}`, this._templateProvider.getRunSettings());
   }
 
-  private async generateSupportFiles(writeFile: WriteFileFn) {
+  private async generateXUnitSupportFiles(writeFile: WriteFileFn) {
     await writeFile(
-      `${StandardOutputFolder.Support}/${"BaseClasses"}${this._outputFileExt}`,
-      (this._templateProvider as PlaywrightCsharpXUnitTemplatesProvider).getBaseClasses(
-        this._rmprojFile.content.rootNamespace,
-        this._rmprojFile.content.testIdAttributeName
-      )
-    );
-    await writeFile(
-      `${StandardOutputFolder.Support}/${StandardOutputFile.TestCaseBase}${this._outputFileExt}`,
-      this._templateProvider.getTestCaseBase(this._rmprojFile.content.rootNamespace)
-    );
-
-    // Filename: Support/TestSuiteBase.cs
-    await writeFile(
-      `${StandardOutputFolder.Support}/${StandardOutputFile.TestSuiteBase}${this._outputFileExt}`,
+      `${StandardOutputFolder.XUnitSupport}/${StandardOutputFile.TestSuiteBase}${this._outputFileExt}`,
       this._templateProvider.getTestSuiteBase(
         this._rmprojFile.content.rootNamespace,
         this._rmprojFile.content.testIdAttributeName
       )
     );
-  }
-
-  private async generateRoutineFiles(writeFile: (path: string, content: string) => Promise<void>) {
-    for (let { content: testRoutine } of this._projMeta.testRoutines) {
-      const datasets: IDataSetInfo[] = generateDatasetInfos(testRoutine);
-      const testRoutinesClasses: string[] = [];
-
-      // For each dataset, we generate a separate routine class
-      for (let dataset of datasets) {
-        let testRoutineClass = this.generateTestRoutineClass(
-          testRoutine,
-          this._projMeta.pages.map((p) => p.content),
-          dataset
-        );
-        testRoutinesClasses.push(testRoutineClass);
-      }
-
-      let testRoutineFile = this.generateTestRoutineFile(testRoutine, testRoutinesClasses);
-      let outputFileRelPath = this._outProjMeta.get(testRoutine.id)!.outputFileRelPath;
-      await writeFile(outputFileRelPath, testRoutineFile);
-    }
   }
 
   private async generateTestSuiteFiles(writeFile: WriteFileFn) {
@@ -132,106 +98,6 @@ export class PlaywrightCsharpXUnitCodeGen extends PlaywrightCsharpCodeGen implem
       );
       await writeFile(fileRelPath, classContent);
     }
-  }
-
-  private async generateTestCaseFiles(writeFile: WriteFileFn) {
-    for (let { content: testCase } of this._projMeta.testCases) {
-      let testClassContent = this.generateTestCaseFile(
-        testCase,
-        this._projMeta.pages.map((p) => p.content),
-        this._projMeta.testRoutines.map((p) => p.content)
-      );
-      let outputFileRelPath = this._outProjMeta.get(testCase.id)!.outputFileRelPath;
-
-      await writeFile(outputFileRelPath, testClassContent);
-    }
-  }
-
-  private async generatePageFiles(writeFile: WriteFileFn) {
-    for (let page of this._projMeta.pages) {
-      let filePath = this._outProjMeta.get(page.content.id)!.outputFileRelPath;
-      await writeFile(filePath, this.generatePage(page.content));
-    }
-  }
-
-  private async generatePageDefinitionsFile(writeFile: WriteFileFn) {
-    const content = this.generatePageDefinitions(this._projMeta.pages.map((p) => p.content));
-    await writeFile(`${StandardOutputFile.PageDefinitions}${this._outputFileExt}`, content);
-  }
-
-  private generatePageDefinitions(pages: IPage[]): string {
-    let usingDirectives: string[] = [];
-    for (let page of pages) {
-      let pageNamespace = this._outProjMeta.get(page.id)!.outputFileFullNamespace;
-      let usingDirective = `using ${pageNamespace};`;
-      if (!usingDirectives.includes(usingDirective)) {
-        usingDirectives.push(usingDirective);
-      }
-    }
-
-    let usings = usingDirectives.join(EOL);
-
-    let pagesDeclarationItems = [];
-    for (let page of pages) {
-      let pageName = upperCaseFirstChar(this._outProjMeta.get(page.id)!.outputFileClassName);
-      pagesDeclarationItems.push(`public ${pageName} ${pageName} { get; private set; }`);
-    }
-    let pagesDeclarations = pagesDeclarationItems.join(EOL);
-    pagesDeclarations = addIndent(pagesDeclarations, this._indentString);
-    //
-    // Build constructor body
-    // Example:
-    // this.LoginPage = new LoginPage(this);
-    //
-    let pageInitItems = [];
-    for (let page of pages) {
-      let pageName = upperCaseFirstChar(this._outProjMeta.get(page.id)!.outputFileClassName);
-      pageInitItems.push(`${pageName} = new ${pageName}(page);`);
-    }
-
-    let pageInits = pageInitItems.join(EOL);
-    pageInits = addIndent(pageInits, this._indentString, 2);
-
-    return this._templateProvider.getPageDefinitions(usings, this._rootNamespace, pagesDeclarations, pageInits);
-  }
-
-  private generatePage(page: IPage): string {
-    let pageItems = [];
-
-    for (let element of page.elements) {
-      if (element.type === "pageElement") {
-        pageItems.push(
-          this._templateProvider.getLocator({
-            elementName: upperCaseFirstChar(element.name!),
-            locatorStr: element.locator || "",
-            locatorType: element.findBy!,
-            description: element.description!,
-            hasParams: hasPlaceholder(element.locator!),
-            returnedLocatorType:
-              element.findBy! === LocatorType.IFrame ||
-              element.findBy! === LocatorType.IFrameId ||
-              element.findBy! === LocatorType.IFrameName
-                ? "IFrameLocator"
-                : "ILocator",
-          })
-        );
-        continue;
-      }
-      if (element.type === "comment") {
-        pageItems.push(""); // Add an empty line before comment
-        pageItems.push(this._templateProvider.getComment(element.comment!));
-      }
-    }
-
-    let pageBody = pageItems.join(EOL + EOL);
-    pageBody = addIndent(pageBody, this._indentString);
-
-    return this._templateProvider.getPage(
-      this._outProjMeta.get(page.id)!.outputFileFullNamespace,
-      this._outProjMeta.get(page.id)!.outputFileClassName,
-      page.description || "",
-      pageBody
-    );
   }
 
   private generateTestSuiteFile(testSuite: ITestSuite, testcases: ITestCase[]) {
@@ -269,51 +135,5 @@ export class PlaywrightCsharpXUnitCodeGen extends PlaywrightCsharpCodeGen implem
       this._outProjMeta.get(testSuite.id)!.outputFileFullNamespace
     );
     return testClass;
-  }
-
-  private generateTestCaseFile(testCase: ITestCase, pages: IPage[], routines: ITestRoutine[]) {
-    const testcaseBody = this.generateTestCaseBody(testCase, pages, routines);
-    const routineUsings = this.generateRoutineUsings(testCase, routines);
-
-    let testFile = this._templateProvider.getTestCaseFile(
-      this._outProjMeta.get(testCase.id)!.outputFileClassName,
-      testCase.description,
-      testcaseBody,
-      this._rootNamespace,
-      this._outProjMeta.get(testCase.id)!.outputFileFullNamespace,
-      routineUsings
-    );
-    return testFile;
-  }
-
-  private generateTestCaseFunction(testCase: ITestCase) {
-    const testcaseName = this._outProjMeta.get(testCase.id)!.outputFileClassName;
-    const testCaseMethod = this._templateProvider.getTestFunction(testcaseName, testCase.description);
-    return testCaseMethod;
-  }
-
-  private generateTestRoutineClass(testRoutine: ITestRoutine, pages: IPage[], datasetInfo: IDataSetInfo) {
-    const testRoutineBody = this.generateTestRoutineBody(testRoutine, pages, datasetInfo);
-
-    // Output class name will be "{testRoutineClassName}{datasetName}";
-    const testRoutineName = this._outProjMeta.get(testRoutine.id)!.outputFileClassName;
-    const finalOutputClassName = `${testRoutineName}${datasetInfo.name}`;
-
-    let routineFileContent = this._templateProvider.getTestRoutineClass(
-      finalOutputClassName,
-      testRoutine.description,
-      testRoutineBody
-    );
-
-    return routineFileContent;
-  }
-
-  private generateTestRoutineFile(testRoutine: ITestRoutine, testRoutineClasses: string[]) {
-    let routineFileContent = this._templateProvider.getTestRoutineFile(
-      this._rootNamespace,
-      this._outProjMeta.get(testRoutine.id)!.outputFileFullNamespace,
-      testRoutineClasses
-    );
-    return routineFileContent;
   }
 }
