+++ common/tests/codegen-tests/playwright-csharp-nunit/playwright-csharp-nunit.test.ts
@@ -9,11 +9,12 @@ import { createTempDir } from "../../test-helpers/fsHelpers";
 import { AutomationFramework, Indent, Language, StandardFolder, TestFramework } from "../../../src/file-defs";
 import { RmpSpec } from "../../test-helpers/rm-project-spec.types";
 import { createTestDataKitchenSink } from "../playwright-typescript/playwright-typescript-kitchen-sink.test-data";
+const PROJ_NAME = "playwright-csharp-nunit";
 
 export const createPlaywrightNunitTestData = (): RmpSpec => {
   const kitchenSinkData = createTestDataKitchenSink();
   return {
-    projectName: "google-test-playwright-nunit",
+    projectName: PROJ_NAME,
     content: {
       fileVersion: 1,
       name: "",
@@ -38,7 +39,7 @@ export const createPlaywrightNunitTestData = (): RmpSpec => {
 
 test("CodeGen Playwright CSharp Nunit - General", async () => {
   // Arrange
-  const tmpDir = createTempDir("playwright-csharp-nunit-general");
+  const tmpDir = createTempDir(PROJ_NAME);
   const projSpec = createPlaywrightNunitTestData();
   const copyToDir = path.join(tmpDir, "rmproj");
   fs.mkdirSync(copyToDir);
